+++ pkg/providers/dns/testdata/TestResDnsZone/update_multisigner.tf
@@ -7,7 +7,7 @@ resource "akamai_dns_zone" "multi_signer_test_zone" {
   zone           = "multisignerexampleterraform.io"
   type           = "primary"
   comment        = "This is an updated test zone with multi-signer DNSSEC"
-  sign_and_serve = false
+  sign_and_serve = true
   group          = "grp1"
 
   multi_provider_dnssec {